rename server/app files and process.exit
