+++ content/webapp/test/server.test.ts
@@ -3,9 +3,12 @@
  */
 import supertest, { SuperTest } from 'supertest';
 import { Server } from 'http';
-import serverPromise from '../start';
+import serverPromise from '../server';
 
 jest.mock('@weco/common/server-data');
+const mockExit = jest.spyOn(process, 'exit').mockImplementation((() => {
+  // never
+}) as () => never);
 
 let server: Server;
 let request: SuperTest<supertest.Test>;
@@ -17,6 +20,7 @@ beforeAll(async () => {
 
 afterAll(() => {
   server.close();
+  expect(mockExit).toHaveBeenCalledWith(0);
 });
 
 test('healthcheck', async () => {
